+++ lib/doorkeeper.rb
@@ -21,6 +21,7 @@ require 'doorkeeper/oauth/code_response'
 require 'doorkeeper/oauth/token_response'
 require 'doorkeeper/oauth/error_response'
 require 'doorkeeper/oauth/pre_authorization'
+require 'doorkeeper/oauth/request_concern'
 require 'doorkeeper/oauth/authorization_code_request'
 require 'doorkeeper/oauth/refresh_token_request'
 require 'doorkeeper/oauth/password_access_token_request'
@@ -30,6 +31,7 @@ require 'doorkeeper/oauth/token_request'
 require 'doorkeeper/oauth/client'
 require 'doorkeeper/oauth/token'
 require 'doorkeeper/oauth/invalid_token_response'
+require 'doorkeeper/oauth/forbidden_token_response'
 
 require 'doorkeeper/models/scopes'
 require 'doorkeeper/models/expirable'
@@ -52,5 +54,9 @@ module Doorkeeper
 
   def self.installed?
     configured? && database_installed?
+  end
+
+  def self.authenticate(request, methods = Doorkeeper.configuration.access_token_methods)
+    OAuth::Token.authenticate(request, *methods)
   end
 end