+++ lib/doorkeeper/config.rb
@@ -35,7 +35,7 @@ module Doorkeeper
 
   def self.setup_application_owner
     require File.join(File.dirname(__FILE__), 'models', 'ownership')
-    Doorkeeper::Application.send :include, Doorkeeper::Models::Ownership
+    Application.send :include, Models::Ownership
   end
 
   class Config
@@ -59,11 +59,11 @@ module Doorkeeper
       end
 
       def default_scopes(*scopes)
-        @config.instance_variable_set('@default_scopes', Doorkeeper::OAuth::Scopes.from_array(scopes))
+        @config.instance_variable_set('@default_scopes', OAuth::Scopes.from_array(scopes))
       end
 
       def optional_scopes(*scopes)
-        @config.instance_variable_set('@optional_scopes', Doorkeeper::OAuth::Scopes.from_array(scopes))
+        @config.instance_variable_set('@optional_scopes', OAuth::Scopes.from_array(scopes))
       end
 
       def client_credentials(*methods)
@@ -80,6 +80,18 @@ module Doorkeeper
 
       def realm(realm)
         @config.instance_variable_set('@realm', realm)
+      end
+
+      def reuse_access_token
+        @config.instance_variable_set("@reuse_access_token", true)
+      end
+
+      def test_redirect_uri(uri)
+        warn <<-TEXT
+          DEPRECATION: test_redirect_uri has renamed to native_redirect_uri. use "native_redirect_uri '#{uri}'".
+        TEXT
+
+        @config.instance_variable_set('@native_redirect_uri', uri)
       end
     end
 
@@ -164,12 +176,14 @@ module Doorkeeper
     option :access_token_expires_in,       default: 7200
     option :authorization_code_expires_in, default: 600
     option :orm,                           default: :active_record
-    option :test_redirect_uri,             default: 'urn:ietf:wg:oauth:2.0:oob'
+    option :native_redirect_uri,           default: 'urn:ietf:wg:oauth:2.0:oob'
     option :active_record_options,         default: {}
     option :realm,                         default: 'Doorkeeper'
     option :wildcard_redirect_uri,         default: false
     option :grant_flows,
            default: %w(authorization_code implicit password client_credentials)
+
+    attr_reader :reuse_access_token
 
     def refresh_token_enabled?
       !!@refresh_token_enabled
@@ -184,11 +198,11 @@ module Doorkeeper
     end
 
     def default_scopes
-      @default_scopes ||= Doorkeeper::OAuth::Scopes.new
+      @default_scopes ||= OAuth::Scopes.new
     end
 
     def optional_scopes
-      @optional_scopes ||= Doorkeeper::OAuth::Scopes.new
+      @optional_scopes ||= OAuth::Scopes.new
     end
 
     def scopes