+++ lib/doorkeeper/models/active_record/access_grant.rb
@@ -6,6 +6,6 @@ module Doorkeeper
 
     serialize :meta
 
-    self.table_name = "#{self.table_name_prefix}oauth_access_grants#{self.table_name_suffix}".to_sym
+    self.table_name = "#{table_name_prefix}oauth_access_grants#{table_name_suffix}".to_sym
   end
 end