+++ lib/doorkeeper/models/mongo_mapper/access_grant.rb
@@ -1,9 +1,6 @@
-require 'doorkeeper/models/mongo_mapper/revocable'
-
 module Doorkeeper
   class AccessGrant
     include MongoMapper::Document
-    include Doorkeeper::Models::MongoMapper::Revocable
     safe
     timestamps!