+++ lib/doorkeeper/oauth/authorization_code_request.rb
@@ -1,7 +1,8 @@
 module Doorkeeper
   module OAuth
     class AuthorizationCodeRequest
-      include Doorkeeper::Validations
+      include Validations
+      include OAuth::RequestConcern
 
       validate :attributes,   error: :invalid_request
       validate :client,       error: :invalid_client
@@ -17,31 +18,15 @@ module Doorkeeper
         @redirect_uri = parameters[:redirect_uri]
       end
 
-      def authorize
-        validate
-        @response = if valid?
-                      grant.revoke
-                      issue_token
-                      TokenResponse.new access_token
-                    else
-                      ErrorResponse.from_request self
-                    end
-      end
-
-      def valid?
-        error.nil?
-      end
-
       private
 
-      def issue_token
-        @access_token = Doorkeeper::AccessToken.create!(
-          application_id:    grant.application_id,
-          resource_owner_id: grant.resource_owner_id,
-          scopes:            grant.scopes_string,
-          expires_in:        server.access_token_expires_in,
-          use_refresh_token: server.refresh_token_enabled?,
-          meta:              grant.meta)
+      def before_successful_response
+        grant.revoke
+        find_or_create_access_token(grant.application,
+                                    grant.resource_owner_id,
+                                    grant.scopes,
+                                    grant.meta,
+                                    server)
       end
 
       def validate_attributes