+++ spec/dummy/db/schema.rb
@@ -22,6 +22,7 @@ ActiveRecord::Schema.define(version: 20130902175349) do
     t.datetime 'created_at',                        null: false
     t.datetime 'revoked_at'
     t.string   'scopes'
+    t.string   'meta'
   end
 
   add_index 'oauth_access_grants', ['token'], name: 'index_oauth_access_grants_on_token', unique: true
@@ -35,6 +36,7 @@ ActiveRecord::Schema.define(version: 20130902175349) do
     t.datetime 'revoked_at'
     t.datetime 'created_at',        null: false
     t.string   'scopes'
+    t.string   'meta'
   end
 
   add_index 'oauth_access_tokens', ['refresh_token'], name: 'index_oauth_access_tokens_on_refresh_token', unique: true