+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## master
+
+## 1.4.0
+
+- internals
+  - [#427] Adds specs expectations.
+  - [#428] Error response refactor.
+  - [#417] Moves token validation into Access Token class.
+  - [#439] Removes redundant module includes.
+  - [#443] TokensController and TokenInfoController inherit from ActionController::Metal
+- bug
+  - [#418] fixes #243, requests with insufficient scope now respond 403 instead
+    of 401. (API change)
+  - [#438] fixes #398, native redirect for implicit token grant bug.
+  - [#440] namespace fixes
+- enhancements
+  - [#432] Keeps query parameters
+
+## 1.3.1
+
+- enhancements
+  - [#405] Adds facade to more easily get the token from a request in a route
+    constraint.
+  - [#415] Extend Doorkeeper TokenResponse with an `after_successful_response`
+    callback that allows handling of `response` object.
+- internals
+  - [#409] Deprecates `test_redirect_uri` in favor of `native_redirect_uri`.
+    See discussion in: [#351].
+  - [#411] Clean rspec deprecations. General test improvements.
+  - [#412] rspec line width can go longer than 80 (hound CI config).
+- bug
+  - [#413] fixes #340, routing scope is now taken into account in redirect.
+  - [#401] and [#425] application is not required any longer for access_token.
+
+## 1.3.0
+
+- enhancements
+  - [#387] Adds reuse_access_token configuration option.
+
 ## 1.2.0
 
 - enhancements